Update ShiftReview.html
Changed the user story wording at the top of the page because it was a carry-over from Software Development class; the update reflects what the Shift Review page is trying to achieve. 
More important changes proposed here are related to the chart which will be generated based on the data in the table (i.e. as the shifts are populated, the chart will show student progress). To do this, I added a new script to run something called Canvas Charts, and I made edits to the table and chart divs.